moved files from src/app/content/popUpElements to src/app/content/elements/popUp
