+++ src/app/content/elements/popUp/ChapterFilter.spec.tsx
@@ -2,20 +2,20 @@ import { HighlightColorEnum } from '@openstax/highlighter/dist/api';
 import React from 'react';
 import { Provider } from 'react-redux';
 import renderer from 'react-test-renderer';
-import createTestStore from '../../../test/createTestStore';
-import { book as archiveBook, page } from '../../../test/mocks/archiveLoader';
-import { mockCmsBook } from '../../../test/mocks/osWebLoader';
-import AllOrNone from '../../components/AllOrNone';
-import Checkbox from '../../components/Checkbox';
-import MessageProvider from '../../MessageProvider';
-import { Store } from '../../types';
-import { assertDefined } from '../../utils';
-import { receiveBook, receivePage } from '../actions';
-import { receiveHighlightsTotalCounts } from '../highlights/actions';
-import { ConnectedChapterFilter } from '../highlights/components/SummaryPopup/Filters';
-import { HighlightLocationFilters } from '../highlights/types';
-import { formatBookData } from '../utils';
-import { findArchiveTreeNode } from '../utils/archiveTreeUtils';
+import createTestStore from '../../../../test/createTestStore';
+import { book as archiveBook, page } from '../../../../test/mocks/archiveLoader';
+import { mockCmsBook } from '../../../../test/mocks/osWebLoader';
+import AllOrNone from '../../../components/AllOrNone';
+import Checkbox from '../../../components/Checkbox';
+import MessageProvider from '../../../MessageProvider';
+import { Store } from '../../../types';
+import { assertDefined } from '../../../utils';
+import { receiveBook, receivePage } from '../../actions';
+import { receiveHighlightsTotalCounts } from '../../highlights/actions';
+import { ConnectedChapterFilter } from '../../highlights/components/SummaryPopup/Filters';
+import { HighlightLocationFilters } from '../../highlights/types';
+import { formatBookData } from '../../utils';
+import { findArchiveTreeNode } from '../../utils/archiveTreeUtils';
 
 describe('ChapterFilter', () => {
   const book = formatBookData(archiveBook, mockCmsBook);
